feat: verify buffer ark script (#278)
* feat: add script to create multi-chain governance proposal for updating SummerToken's OApp delegate

This script allows users to create a governance proposal that updates the OApp delegate for SummerToken across multiple chains, ensuring that the delegate is set to the Timelock Controller. It includes interactive prompts for configuration selection and confirmation of the proposal details before execution.

* chore: remove update-oapp-delegates script and add buffer-arks verification script

This commit deletes the `update-oapp-delegates.ts` script, which was responsible for creating multi-chain governance proposals for updating SummerToken's OApp delegate. In its place, a new script `buffer-arks.ts` is introduced to verify buffer ARK deployments across specified networks, including interactive configuration prompts and detailed logging for verification processes.

* Update Codecov badge to current branch

---------

Co-authored-by: GitHub Action <action@github.com>